Removes old reference code
Adds a section for generalizing the mission track
The code at this stage is not organized. I need to separate the function
PlotMissionTrack into 2 functions. One function to get the track
and the other is for plotting the track. Plotting the track is only
for exploration and is not included in the metadata